fix: swap issue resolve
