+++ package.json
@@ -20,20 +20,20 @@
     "@ethersproject/strings": "5.6.1",
     "@ethersproject/wallet": "5.6.2",
     "@gnosis.pm/safe-apps-web3-react": "0.6.8",
-    "@honeycomb-finance/airdrop": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/bridge": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/core": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/elixir": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/governance": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/honeycomb-provider": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/portfolio": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/pools": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/sar": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/shared": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/state-hooks": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/swap": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/wallet-connectors": "0.0.1-internal-20230921064423",
-    "@honeycomb-finance/walletmodal": "0.0.1-internal-20230921064423",
+    "@honeycomb-finance/airdrop": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/bridge": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/core": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/elixir": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/governance": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/honeycomb-provider": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/pools": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/portfolio": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/sar": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/shared": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/state-hooks": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/swap": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/wallet-connectors": "0.0.1-internal-20230922074227",
+    "@honeycomb-finance/walletmodal": "0.0.1-internal-20230922074227",
     "@metamask/jazzicon": "2.0.0",
     "@pangolindex/components": "7.3.10-rc.0",
     "@pangolindex/exchange-contracts": "2.1.7",
@@ -99,7 +99,7 @@
     "react-error-overlay": "6.0.9",
     "react-feather": "2.0.9",
     "react-markdown": "8.0.1",
-    "react-query": "3.34.16",
+    "react-query": "3.39.3",
     "react-redux": "7.2.6",
     "react-router-dom": "6.8.1",
     "react-spring": "8.0.27",
@@ -111,6 +111,7 @@
     "recharts": "2.1.13",
     "redux-localstorage-simple": "2.4.1",
     "remark-gfm": "3.0.1",
+    "rollup-plugin-node-polyfills": "0.2.1",
     "serve": "11.3.2",
     "start-server-and-test": "1.14.0",
     "styled-components": "5.3.1",
