+++ src/components/AccountDetails/index.tsx
@@ -8,7 +8,7 @@ import Copy from './Copy'
 import Transaction from './Transaction'
 import { useAllTransactionsClearer } from '@honeycomb-finance/state-hooks'
 import { NearConnector, injected } from '@honeycomb-finance/wallet-connectors'
-import { PangolinInjectedWallet } from '@honeycomb-finance/walletmodal'
+import { HoneycombInjectedWallet } from '@honeycomb-finance/walletmodal'
 import { useActiveWeb3React, useTranslation, getEtherscanLink, shortenAddressMapping } from '@honeycomb-finance/shared'
 import Identicon from '../Identicon'
 import { ExternalLink as LinkIcon } from 'react-feather'
@@ -70,7 +70,7 @@ export default function AccountDetails({
   const wallet = Object.values(supportedWallets).find(_wallet => _wallet.isActive)
 
   function getStatusIcon() {
-    if (wallet instanceof PangolinInjectedWallet || wallet?.connector === injected) {
+    if (wallet instanceof HoneycombInjectedWallet || wallet?.connector === injected) {
       return (
         <IconWrapper size={16}>
           <Identicon />
