+++ src/pages/Beta/Swap/index.tsx
@@ -1,65 +1,29 @@
-import React, { useState } from 'react'
-import { PageWrapper, GridContainer, TopContainer, StatsWrapper, GridWrapper } from './styleds'
-import { SwapWidget, SwapTypes, useGelatoLimitOrdersListHook } from '@honeycomb-finance/swap'
-import { WatchList, MyPortfolio } from '@honeycomb-finance/portfolio'
-import PairInfo from './PairInfo'
-import LimitOrderList from './LimitOrderList'
+import React from 'react'
+import SwapUI from './SwapUI'
+import { useLibrary, useActiveWeb3React } from '@honeycomb-finance/shared'
 import { useChainId } from 'src/hooks'
-import { CHAINS } from '@pangolindex/sdk'
-import ComingSoon from 'src/components/Beta/ComingSoon'
-import { LIMITORDERLIST_ACCESS, MYPORTFOLIO_ACCESS, PAIRINFO_ACCESS } from 'src/constants/accessPermissions'
-import { Hidden, Visible } from 'src/theme'
+import { GelatoProvider, galetoStore } from '@honeycomb-finance/swap'
+import { Web3Provider } from '@ethersproject/providers'
+import { Provider } from 'react-redux'
 
-const SwapUI = () => {
+const Swap = () => {
   const chainId = useChainId()
-  const useGelatoLimitOrders = useGelatoLimitOrdersListHook[chainId]
-  const { allOrders } = useGelatoLimitOrders()
-  const [swapType, onSwapTypeChange] = useState(SwapTypes.MARKET)
-  const isLimitOrders = (allOrders || []).length > 0 && swapType === SwapTypes.LIMIT
+  const { account } = useActiveWeb3React()
+  const { library } = useLibrary()
 
-  const isOnlySwapWidget = !PAIRINFO_ACCESS[chainId] && !LIMITORDERLIST_ACCESS[chainId] && !MYPORTFOLIO_ACCESS[chainId]
-
-  const getSwapPageInfo = () => {
-    if (isOnlySwapWidget) {
-      return (
-        <GridWrapper>
-          <Hidden upToSmall={true}>
-            <WatchList coinChartVisible={true} />
-          </Hidden>
-          <SwapWidget
-            isTWAPOrderVisible={CHAINS[chainId]?.mainnet && CHAINS[chainId]?.supported_by_twap}
-            onSwapTypeChange={onSwapTypeChange}
-            isLimitOrderVisible={CHAINS[chainId]?.mainnet && CHAINS[chainId]?.supported_by_gelato}
-          />
-          <Visible upToSmall={true} upToMedium={false}>
-            <WatchList coinChartVisible={true} />
-          </Visible>
-        </GridWrapper>
-      )
-    } else {
-      return (
-        <>
-          <TopContainer>
-            <StatsWrapper>{PAIRINFO_ACCESS[chainId] ? <PairInfo /> : <ComingSoon />}</StatsWrapper>
-            <SwapWidget
-              isTWAPOrderVisible={CHAINS[chainId]?.mainnet && CHAINS[chainId]?.supported_by_twap}
-              onSwapTypeChange={onSwapTypeChange}
-              isLimitOrderVisible={CHAINS[chainId]?.mainnet && CHAINS[chainId]?.supported_by_gelato}
-            />
-          </TopContainer>
-
-          {CHAINS[chainId]?.mainnet && (
-            <GridContainer isLimitOrders={isLimitOrders}>
-              {isLimitOrders && LIMITORDERLIST_ACCESS[chainId] && <LimitOrderList />}
-              {MYPORTFOLIO_ACCESS[chainId] && <MyPortfolio />}
-              <WatchList coinChartVisible={!isLimitOrders} />
-            </GridContainer>
-          )}
-        </>
-      )
-    }
-  }
-
-  return <PageWrapper>{getSwapPageInfo()}</PageWrapper>
+  const ethersLibrary = library && !library?._isProvider ? new Web3Provider(library) : library
+  return (
+    <Provider store={galetoStore}>
+      <GelatoProvider
+        library={ethersLibrary}
+        chainId={chainId}
+        account={account ?? undefined}
+        useDefaultTheme={false}
+        handler={'pangolin'}
+      >
+        <SwapUI />
+      </GelatoProvider>
+    </Provider>
+  )
 }
-export default SwapUI
+export default Swap
